+++ src/runtime.rs
@@ -91,7 +91,7 @@ impl Runtime for DOM {
          
             if self.control.is_some() && self.control.as_ref().unwrap().is_loader_busy() {
                 let control = self.control.as_mut().unwrap();
-                self.surface.draw_progress_screen(control.get_loader_progress(255).take().unwrap().try_into().unwrap(), control.get_loader_error());            
+                self.surface.draw_progress_screen(control.get_loader_progress(255).take().unwrap().try_into().unwrap(), control.get_loader_error() as u8);            
             }else{
                 self.surface.clear(0);
                 self.surface.update(&mut self.dom);
